realtime/requirements.txt
python-dotenv==1.0.1
pvporcupine==3.0.5
pvrecorder==1.2.5
websocket-client==1.8.0
